backend/buddy_guyAI_api/requirements.txt
boto3==1.17.58
Django==3.2.2
django-cors-headers==3.7.0
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-yasg==1.20.0
gunicorn==20.1.0
Pillow==8.2.0
pandas==1.2.4
psycopg2-binary==2.8.6
python-dotenv==0.18.0
requests==2.25.1
scikit-learn==0.24.2
scipy==1.7.1
wavio==0.0.4